requirements.txt
Flask
pytz
firebase_admin
bs4
flask-cors
google-cloud-firestore
python-dotenv
psycopg2-binary
